+++ routes/api/contacts.js
@@ -1,86 +1,31 @@
 const express = require("express");
-const Joi = require("joi");
+// const Joi = require("joi");
 const router = express.Router();
+const ctrl = require("../../controllers/contacts")
 
-const contacts = require("../../models/contacts");
-const { HttpError } = require("../../helpers");
-const { addContact } = require("../../models/contacts");
-
-
-const addShema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
-
-router.get("/", async (req, res, next) => {
-  try {
-    const results = await contacts.listContacts();
-    res.json(results);
-  } catch (error) {
-    next(error);
-  }
-});
+// const contacts = require("../../models/contacts");
+// const { HttpError } = require("../../helpers");
+// const { addContact } = require("../../models/contacts");
 
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
+// const addShema = Joi.object({
+//   name: Joi.string().required(),
+//   email: Joi.string().required(),
+//   phone: Joi.string().required(),
+// });
 
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", ctrl.getAll );
 
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addShema.validate(req.body);
-    if (error) {
-      throw HttpError(error, error.message);
-    }
-    const result = await addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", ctrl.getById);
 
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json({ message: "Delete success" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", ctrl.add);
 
 
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = addShema.validate(req.body);
-    if (error) {
-      throw HttpError(error, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", ctrl.deleteById);
+
+
+router.put("/:contactId", ctrl.updateById);
 
 module.exports = router;
